download only mode

+++ src/operations/package-manager/remove.vala
@@ -39,22 +39,26 @@ public int remove_main(string[] args){
 
 public int remove_single(package p){
     print(colorize(_("Removing: %s"),yellow).printf(p.name));
-    foreach(string file in p.list_files()){
-        if(file.length > 41){
-            file=file[41:];
-            remove_file(DESTDIR+"/"+file);
-        }
-    }foreach(string file in p.list_links()){
-        if(file.length > 3){
-            file=ssplit(file," ")[0];
-            remove_file(DESTDIR+"/"+file);
+    if(!get_bool("without-files")){
+        foreach(string file in p.list_files()){
+            if(file.length > 41){
+                file=file[41:];
+                remove_file(DESTDIR+"/"+file);
+            }
+        }foreach(string file in p.list_links()){
+            if(file.length > 3){
+                file=ssplit(file," ")[0];
+                remove_file(DESTDIR+"/"+file);
+            }
         }
     }
-    remove_file(get_storage()+"/metadata/"+p.name+".yaml");
-    remove_file(get_storage()+"/files/"+p.name);
-    remove_file(get_storage()+"/links/"+p.name);
-    if(isdir(get_storage()+"/sysconf/"+p.name)){
-        remove_all(get_storage()+"/sysconf/"+p.name);
+    if(!get_bool("without-metadata")){
+        remove_file(get_storage()+"/metadata/"+p.name+".yaml");
+        remove_file(get_storage()+"/files/"+p.name);
+        remove_file(get_storage()+"/links/"+p.name);
+        if(isdir(get_storage()+"/sysconf/"+p.name)){
+            remove_all(get_storage()+"/sysconf/"+p.name);
+        }
     }
     return 0;
 }
@@ -68,5 +72,7 @@ void remove_init(){
     op.help.minargs=1;
     op.help.description = _("Remove package from package name.");
     op.help.add_parameter("--ignore-dependency", _("disable dependency check"));
+    op.help.add_parameter("--without-files", _("do not remove files from package database"));
+    op.help.add_parameter("--without-metadata", _("do not remove metadata from package database"));
     add_operation(op);
 }
